# web/clase-04/live-middleware-y-documentacion/docs/swagger.yaml
definitions:
  domain.Producto:
    properties:
      code_value:
        type: string
      expiration:
        type: string
      id:
        type: integer
      is_published:
        type: boolean
      name:
        type: string
      price:
        type: number
      quantity:
        type: integer
    type: object
  web.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
  /products:
    get:
      description: Get all products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: List all Products
      tags:
      - Products
    post:
      description: Create product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Producto'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Save Products
      tags:
      - Products
  /products/:id:
    delete:
      description: Delete products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Delete products
      tags:
      - Products
    get:
      description: Get product by id
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: List Products by Id
      tags:
      - Products
    patch:
      description: Update partial product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Producto'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Update partial products
      tags:
      - Products
    put:
      description: Update product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Producto'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      summary: Update products
      tags:
      - Products
swagger: "2.0"